+++ lib/ioki/model/passenger/logpay_url.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Ioki
+  module Model
+    module Passenger
+      class LogpayUrl < Base
+        attribute :url, on: :create, type: :string
+      end
+    end
+  end
+end